requirements.txt
affine==2.3.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.0.0
cycler==0.11.0
Cython==0.29.28
dask==2022.5.0
fonttools==4.33.3
fsspec==2022.3.0
idna==3.3
image-similarity-measures==0.3.5
imageio==2.19.0
imutils==0.5.4
kiwisolver==1.4.2
locket==1.0.0
matplotlib==3.5.2
networkx==2.8
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
partd==1.2.0
pdf2image==1.16.0
phasepack==1.5
Pillow==9.1.0
piq==0.7.0
pyFFTW==0.13.0
pyparsing==3.0.8
pyssim==0.4
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
rasterio==1.2.10
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.0
sewar==0.4.5
six==1.16.0
snuggs==1.4.7
tifffile==2022.5.4
toolz==0.11.2
torch==1.11.0
torchvision==0.12.0
typing-extensions==4.2.0
urllib3==1.26.9
